requirements.txt
absl-py==2.1.0
ale-py==0.8.1
astor==0.8.1
atari-py==0.2.9
certifi
cffi==1.15.1
charset-normalizer==3.3.2
cloudpickle==2.2.1
cycler==0.11.0
Cython==3.0.0a10
fasteners==0.19
fonttools==4.38.0
gast==0.2.2
glfw==2.6.5
google-pasta==0.2.0
grpcio==1.60.0
gym==0.10.9
gym-notices==0.0.8
h5py==2.10.0
idna==3.6
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.3.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.4
matplotlib==3.5.3
mpi4py
mujoco-py==2.1.2.14
numpy==1.18.5
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
protobuf==3.20.0
pyaml==23.5.8
pycparser==2.21
pygame==2.1.0
pyglet==2.0.10
PyOpenGL==3.1.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==3.13
requests==2.31.0
scipy==1.7.3
six==1.16.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.5
termcolor==2.3.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
